requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.31.0
python-multipart==0.0.9

# Data processing
pandas==2.2.0
openpyxl==3.1.2

# Web scraping
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0

# ML & Embeddings
sentence-transformers==2.3.1
scikit-learn==1.4.0
numpy==1.26.3

# LLM
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.1
pydantic==2.6.0
